Split `RtpStream.ts` into `rtpStreamStatsTypes.ts` and `rtpStreamStatsFbsUtils` Rename `scalabilityModes.ts` tp `scalabilityModesTypes.ts`
