+++ node/src/rtpStreamStatsFbsUtils.ts
@@ -1,44 +1,11 @@
+import {
+	RtpStreamRecvStats,
+	RtpStreamSendStats,
+	BaseRtpStreamStats,
+	BitrateByLayer,
+} from './rtpStreamStatsTypes';
 import * as FbsRtpStream from './fbs/rtp-stream';
 import * as FbsRtpParameters from './fbs/rtp-parameters';
-
-type BitrateByLayer = { [key: string]: number };
-
-export type RtpStreamRecvStats = BaseRtpStreamStats & {
-	type: string;
-	jitter: number;
-	packetCount: number;
-	byteCount: number;
-	bitrate: number;
-	bitrateByLayer: BitrateByLayer;
-};
-
-export type RtpStreamSendStats = BaseRtpStreamStats & {
-	type: string;
-	packetCount: number;
-	byteCount: number;
-	bitrate: number;
-};
-
-type BaseRtpStreamStats = {
-	timestamp: number;
-	ssrc: number;
-	rtxSsrc?: number;
-	rid?: string;
-	kind: string;
-	mimeType: string;
-	packetsLost: number;
-	fractionLost: number;
-	packetsDiscarded: number;
-	packetsRetransmitted: number;
-	packetsRepaired: number;
-	nackCount: number;
-	nackPacketCount: number;
-	pliCount: number;
-	firCount: number;
-	score: number;
-	roundTripTime?: number;
-	rtxPacketsDiscarded?: number;
-};
 
 export function parseRtpStreamStats(
 	binary: FbsRtpStream.Stats
