+++ node/src/scalabilityModesTypes.ts
@@ -1,0 +1,5 @@
+export type ScalabilityMode = {
+	spatialLayers: number;
+	temporalLayers: number;
+	ksvc: boolean;
+};
